requirements.txt
gradio_client==1.10.1
gradio==5.30.0
Pillow==11.0.0
playsound==1.3.0
protobuf==5.29.2
pydub==0.25.1
pygame==2.6.1
pyttsx3==2.98
SpeechRecognition==3.12.0
transformers==4.47.1
torch==2.5.1
google-generativeai